refactor: initialize MainMenu in Core.initializeGUIimpl
- unbundle a menu object from MainWindow class for unit test
