+++ src/org/omegat/core/Core.java
@@ -26,6 +26,11 @@
 
 package org.omegat.core;
 
+import java.awt.BorderLayout;
+import java.awt.Component;
+import java.awt.event.WindowAdapter;
+import java.awt.event.WindowEvent;
+import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -33,6 +38,8 @@ import java.util.Map;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import java.util.concurrent.locks.ReentrantLock;
+
+import javax.swing.UIManager;
 
 import org.omegat.core.data.EntryKey;
 import org.omegat.core.data.IProject;
@@ -65,9 +72,13 @@ import org.omegat.gui.glossary.GlossaryTextArea;
 import org.omegat.gui.glossary.IGlossaries;
 import org.omegat.gui.issues.IIssues;
 import org.omegat.gui.issues.IssuesPanelController;
+import org.omegat.gui.main.BaseMainWindowMenu;
 import org.omegat.gui.main.ConsoleWindow;
+import org.omegat.gui.main.IMainMenu;
 import org.omegat.gui.main.IMainWindow;
 import org.omegat.gui.main.MainWindow;
+import org.omegat.gui.main.MainWindowMenu;
+import org.omegat.gui.main.MainWindowMenuHandler;
 import org.omegat.gui.matches.IMatcher;
 import org.omegat.gui.matches.MatchesTextArea;
 import org.omegat.gui.multtrans.MultipleTransPane;
@@ -119,6 +130,7 @@ public final class Core {
     private static INotes notes;
     private static IComments comments;
     private static Segmenter segmenter;
+    private static BaseMainWindowMenu menu;
 
     private static Map<String, String> cmdLineParams = Collections.emptyMap();
 
@@ -140,6 +152,10 @@ public final class Core {
     /** Get main window instance. */
     public static IMainWindow getMainWindow() {
         return mainWindow;
+    }
+
+    public static IMainMenu getMainMenu() {
+        return menu;
     }
 
     /** Get editor instance. */
@@ -266,6 +282,7 @@ public final class Core {
      * @throws Exception
      */
     static void initializeGUIimpl(MainWindow me) throws Exception {
+        initMainMenu(me);
         MarkerController.init();
         LanguageToolWrapper.init();
 
@@ -297,6 +314,54 @@ public final class Core {
         tagValidation = new TagValidationTool();
         currentProject = new NotLoadedProject();
         mainWindow = new ConsoleWindow();
+    }
+
+    @SuppressWarnings("unchecked")
+    private static void initMainMenu(MainWindow mw) {
+        MainWindowMenuHandler mainWindowMenuHandler = new MainWindowMenuHandler(mw);
+
+        // Load Menu extension
+        Object menuClass = UIManager.get(UIDesignManager.menuClassID);
+        if (menuClass != null) {
+            BaseMainWindowMenu menu1;
+            try {
+                menu1 = ((Class<? extends BaseMainWindowMenu>) menuClass)
+                        .getDeclaredConstructor(MainWindow.class, MainWindowMenuHandler.class)
+                        .newInstance(mw, mainWindowMenuHandler);
+            } catch (Exception e) {
+                // fall back to default when loading failed.
+                menu1 = new MainWindowMenu(mw, mainWindowMenuHandler);
+            }
+            menu = menu1;
+        } else {
+            // Default menu.
+            menu = new MainWindowMenu(mw, mainWindowMenuHandler);
+        }
+        mw.getApplicationFrame().setJMenuBar(menu.getMainMenu());
+
+        mw.getApplicationFrame().addWindowListener(new WindowAdapter() {
+            public void windowClosing(WindowEvent e) {
+                mainWindowMenuHandler.projectExitMenuItemActionPerformed();
+            }
+
+            @Override
+            public void windowDeactivated(WindowEvent we) {
+                Core.getEditor().windowDeactivated();
+            }
+        });
+
+        // Load toolbar extension
+        Object toolbarClass = UIManager.get(UIDesignManager.toolbarClassID);
+        if (toolbarClass != null) {
+            try {
+                mw.getApplicationFrame().getContentPane()
+                        .add((Component) ((Class<?>) toolbarClass)
+                                .getDeclaredConstructor(MainWindow.class, MainWindowMenuHandler.class)
+                                .newInstance(mw, mainWindowMenuHandler), BorderLayout.NORTH);
+            } catch (InstantiationException | IllegalAccessException | InvocationTargetException
+                     | NoSuchMethodException ignored) {
+            }
+        }
     }
 
     /**
